Move DateCell to own component
